scrapers/requirements.txt
playwright
beautifulsoup4
requests
supabase
pdfplumber
python-dateutil
psycopg2-binary
python-dotenv
